revert(main): restaurar árvore exata do commit 9362503 (preserva histórico)
