--- MANIFEST.in
@@ -1,5 +1,0 @@
-include README.md
-include requirements.txt
-include pytest.ini
-recursive-include src *.py
-recursive-include sources *.py
