+++ output/history/motorsport_events_20250803_20250803_220421.ics
@@ -1,0 +1,237 @@
+BEGIN:VCALENDAR
+VERSION:2.0
+PRODID:-//Motorsport Calendar Generator//EN
+CALSCALE:GREGORIAN
+METHOD:PUBLISH
+X-WR-CALDESC:Weekend motorsport events automatically collected
+X-WR-CALNAME:Motorsport Events
+X-APPLE-CALENDAR-COLOR:#FF6B35
+X-WR-TIMEZONE:America/Sao_Paulo
+BEGIN:VEVENT
+SUMMARY:WRC - PROGRAMAÇÃO DA TV E INTERNET
+DTSTART;TZID=America/Sao_Paulo:20250803T120000
+DTEND;TZID=America/Sao_Paulo:20250803T140000
+DTSTAMP:20250803T003235Z
+UID:7e957f6e1e4dfdea@motorsport-calendar
+CATEGORIES:WRC
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 WRC Race\n📍 Brazil\n\n📊 Source: Tomada de Tempo\n\n
+ ⚠️ Category detection confidence: 70%
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:5
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:WRC
+X-MOTORSPORT-SESSION:Race
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:F3 - Hungria de F
+DTSTART;TZID=America/Sao_Paulo:20250803T120000
+DTEND;TZID=America/Sao_Paulo:20250803T140000
+DTSTAMP:20250803T003235Z
+UID:363fd29fdc123bfc@motorsport-calendar
+CATEGORIES:F3
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 F3 Race\n📍 Brazil\n\n📊 Source: Tomada de Tempo
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:4
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:F3
+X-MOTORSPORT-SESSION:Race
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:F2 - Hungria de F
+DTSTART;TZID=America/Sao_Paulo:20250803T120000
+DTEND;TZID=America/Sao_Paulo:20250803T140000
+DTSTAMP:20250803T003235Z
+UID:363fd29fdc123bfc@motorsport-calendar
+CATEGORIES:F2
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 F2 Race\n📍 Brazil\n\n📊 Source: Tomada de Tempo
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:3
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:F2
+X-MOTORSPORT-SESSION:Race
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:F1 - Hungria de F
+DTSTART;TZID=America/Sao_Paulo:20250803T120000
+DTEND;TZID=America/Sao_Paulo:20250803T140000
+DTSTAMP:20250803T003235Z
+UID:363fd29fdc123bfc@motorsport-calendar
+CATEGORIES:F1
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 F1 Race\n📍 Brazil\n\n📊 Source: Tomada de Tempo
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:1
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:F1
+X-MOTORSPORT-SESSION:Race
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:F2 - E tivemos\, neste sábado (02/08/2025)\, (Practice)
+DTSTART;TZID=America/Sao_Paulo:20250802T120000
+DTEND;TZID=America/Sao_Paulo:20250802T133000
+DTSTAMP:20250803T003235Z
+UID:fb2ef06a21068852@motorsport-calendar
+CATEGORIES:F2,Practice
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 F2 Practice\n📍 Brazil\n\n📊 Source: Tomada de Tempo\
+ n\n⚠️ Category detection confidence: 67%
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:3
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:F2
+X-MOTORSPORT-SESSION:Practice
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:F1 - Hungria da F
+DTSTART;TZID=America/Sao_Paulo:20250802T120000
+DTEND;TZID=America/Sao_Paulo:20250802T140000
+DTSTAMP:20250803T003235Z
+UID:ac6441399021c289@motorsport-calendar
+CATEGORIES:F1
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 F1 Race\n📍 Brazil\n\n📊 Source: Tomada de Tempo
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:1
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:F1
+X-MOTORSPORT-SESSION:Race
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:FormulaE - Hungria de (Practice)
+DTSTART;TZID=America/Sao_Paulo:20250801T120000
+DTEND;TZID=America/Sao_Paulo:20250801T133000
+DTSTAMP:20250803T003235Z
+UID:ce9992ce8689c262@motorsport-calendar
+CATEGORIES:FormulaE,Practice
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 FormulaE Practice\n📍 Brazil\n\n📊 Source: Tomada de 
+ Tempo
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:5
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:FormulaE
+X-MOTORSPORT-SESSION:Practice
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+BEGIN:VEVENT
+SUMMARY:F1 - Hungria de F (Practice)
+DTSTART;TZID=America/Sao_Paulo:20250801T120000
+DTEND;TZID=America/Sao_Paulo:20250801T133000
+DTSTAMP:20250803T003235Z
+UID:03ecf7ed24623dcf@motorsport-calendar
+CATEGORIES:F1,Practice
+CREATED:20250803T003235Z
+DESCRIPTION:🏁 F1 Practice\n📍 Brazil\n\n📊 Source: Tomada de Tempo
+LAST-MODIFIED:20250803T003235Z
+LOCATION:Brazil
+PRIORITY:1
+STATUS:CONFIRMED
+TRANSP:OPAQUE
+X-MOTORSPORT-CATEGORY:F1
+X-MOTORSPORT-SESSION:Practice
+X-MOTORSPORT-SOURCE:tomadatempo
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT1H
+END:VALARM
+BEGIN:VALARM
+ACTION:DISPLAY
+DESCRIPTION:Motorsport Event Reminder
+TRIGGER:-PT15M
+END:VALARM
+END:VEVENT
+END:VCALENDAR
