--- pyproject.toml
@@ -1,91 +1,0 @@
-[build-system]
-requires = ["setuptools>=42", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "motorsport-calendar"
-version = "0.1.0"
-description = "A calendar application for motorsport events"
-readme = "README.md"
-authors = [
-    {name = "Daniel Mirrha", email = "dmirrha@gmail.com"}
-]
-license = {text = "MIT"}
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent"
-]
-requires-python = ">=3.8"
-dependencies = [
-    "requests>=2.31.0",
-    "beautifulsoup4>=4.12.0",
-    "pytz>=2023.3",
-    "python-dateutil>=2.8.2",
-    "icalendar>=5.0.0",
-    "python-levenshtein>=0.21.0",
-    "scikit-learn>=1.3.0",
-    "nltk>=3.8.0",
-    "rich>=13.5.0",
-    "colorama>=0.4.6",
-    "tqdm>=4.66.0",
-    "colorlog>=6.7.0",
-    "click>=8.1.0",
-    "pyyaml>=6.0.0"
-]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.4.0",
-    "pytest-cov>=4.1.0",
-    "black>=23.7.0",
-    "flake8>=6.0.0",
-    "freezegun>=1.2.2"
-]
-
-[tool.setuptools]
-package-dir = { "" = "src" }
-packages.find = { where = ["src"] }
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-python_files = ["test_*.py"]
-addopts = [
-    "-v",
-    "--cov=src",
-    "--cov=sources",
-    "--cov-report=term-missing",
-    "--cov-report=xml:test_results/coverage.xml",
-    "--junitxml=test_results/junit.xml",
-    "--html=test_results/report.html",
-    "--self-contained-html"
-]
-
-[tool.coverage.run]
-source = ["src", "sources"]
-branch = true
-
-[tool.coverage.report]
-show_missing = true
-skip_covered = true
-fail_under = 80
-
-[tool.black]
-line-length = 120
-target-version = ["py38"]
-include = '\.pyi?$'
-
-[tool.flake8]
-max-line-length = 120
-exclude = [
-    ".git",
-    "__pycache__",
-    "build",
-    "dist",
-    ".venv",
-    "venv"
-]
-per-file-ignores = [
-    "__init__.py: F401",
-    "tests/*.py: E402"
-]
