--- pytest.ini
@@ -1,34 +1,0 @@
-[pytest]
-# Test discovery configuration
-testpaths = tests
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-
-# Test execution options
-addopts = 
-    -v 
-    --cov=src 
-    --cov=sources
-    --cov-report=term-missing 
-    --cov-report=xml:test_results/coverage.xml
-    --junitxml=test_results/junit.xml
-    --html=test_results/report.html
-    --self-contained-html
-
-# Logging configuration
-log_cli = true
-log_cli_level = INFO
-log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
-log_cli_date_format = %Y-%m-%d %H:%M:%S
-
-log_file = test_results/pytest.log
-log_file_level = DEBUG
-log_file_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
-log_file_date_format = %Y-%m-%d %H:%M:%S
-
-# Test markers
-markers =
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    integration: mark tests that require external services
-    regression: regression tests
