+++ requirements.txt
@@ -42,7 +42,6 @@ pytest>=7.4.0
 pytest-cov>=4.1.0
 black>=23.7.0
 flake8>=6.0.0
-freezegun>=1.2.2
 
 # System and File Operations (Built-in, listed for reference)
 # pathlib - Built-in Python 3.8+
