--- setup.cfg
@@ -1,38 +1,0 @@
-[metadata]
-description-file = README.md
-
-[tool:pytest]
-testpaths = tests
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-addopts = 
-    -v 
-    --cov=src 
-    --cov=sources
-    --cov-report=term-missing 
-    --cov-report=xml:test_results/coverage.xml
-    --junitxml=test_results/junit.xml
-    --html=test_results/report.html
-    --self-contained-html
-
-[coverage:run]
-source = src,sources
-branch = True
-
-[coverage:report]
-show_missing = True
-skip_covered = True
-fail_under = 80
-
-[coverage:paths]
-source =
-    src/
-    */site-packages/
-
-[flake8]
-max-line-length = 120
-exclude = .git,__pycache__,build,dist,.venv,venv
-per-file-ignores =
-    __init__.py: F401
-    tests/*.py: E402
