+++ sources/tomada_tempo.py
@@ -6,63 +6,15 @@ Brazilian motorsport calendar website with complete schedules and streaming info
 """
 
 import re
-import time
-import random
-import requests
 from datetime import datetime, timedelta
-from typing import List, Dict, Any, Optional, Tuple, Union
-from bs4 import BeautifulSoup, Tag
+from typing import List, Dict, Any, Optional
+from bs4 import BeautifulSoup
 from urllib.parse import urljoin, urlparse
 from .base_source import BaseSource
-from typing import Optional, Tuple
 
 
 class TomadaTempoSource(BaseSource):
     """Primary data source for tomadadetempo.com.br"""
-    
-    def _parse_event_time_range(self, time_range: str, base_date: datetime) -> Tuple[Optional[datetime], Optional[datetime]]:
-        """Parse a time range string into start and end datetimes.
-        
-        Args:
-            time_range: Time range string (e.g., "14:00 - 16:00")
-            base_date: Base date to use for the time
-            
-        Returns:
-            Tuple of (start_time, end_time) or (None, None) if parsing fails
-        """
-        try:
-            if not time_range or not isinstance(time_range, str):
-                return None, None
-                
-            # Handle time range format: "14:00 - 16:00"
-            if ' - ' in time_range:
-                start_str, end_str = time_range.split(' - ', 1)
-                start_time = datetime.strptime(start_str.strip(), '%H:%M').time()
-                end_time = datetime.strptime(end_str.strip(), '%H:%M').time()
-                
-                # Combine with base date
-                start_dt = datetime.combine(base_date, start_time)
-                end_dt = datetime.combine(base_date, end_time)
-                
-                # Handle overnight events
-                if end_time < start_time:
-                    end_dt += timedelta(days=1)
-                    
-                return start_dt, end_dt
-                
-            # Handle single time format: "14:00"
-            elif ':' in time_range:
-                time_obj = datetime.strptime(time_range.strip(), '%H:%M').time()
-                start_dt = datetime.combine(base_date, time_obj)
-                # Default duration of 2 hours if no end time specified
-                end_dt = start_dt + timedelta(hours=2)
-                return start_dt, end_dt
-                
-        except Exception as e:
-            if self.logger:
-                self.logger.warning(f"Failed to parse time range '{time_range}': {str(e)}")
-                
-        return None, None
     
     def get_display_name(self) -> str:
         """Get human-readable display name."""
@@ -71,419 +23,91 @@ class TomadaTempoSource(BaseSource):
     def get_base_url(self) -> str:
         """Get base URL for Tomada de Tempo."""
         return "https://www.tomadadetempo.com.br"
-        
+    
     def collect_events(self, target_date: Optional[datetime] = None) -> List[Dict[str, Any]]:
         """
         Collect motorsport events from Tomada de Tempo.
         
         Args:
-            target_date: Target date for event collection (defaults to current date)
+            target_date: Target date for event collection
             
         Returns:
-            List of event dictionaries with standardized format
-            
-        Raises:
-            Exception: If there's an error during event collection
+            List of event dictionaries
         """
-        if target_date is None:
-            target_date = datetime.now()
-            
         if self.logger:
-            self.logger.info(f"🔍 Collecting events from {self.get_display_name()} for {target_date.strftime('%Y-%m-%d')}")
+            self.logger.log_source_start(self.source_display_name)
+        
+        if self.ui:
+            self.ui.show_source_result(self.source_display_name, True, 0, "Starting collection...")
+        
+        events = []
         
         try:
-            # First try to get events from weekend programming
-            events = self._collect_from_weekend_programming(target_date)
+            # Get current weekend if no target date provided
+            if not target_date:
+                target_date = self._get_next_weekend()
             
-            # If no events found in weekend programming, try the general calendar
+            # Calculate current weekend range (Friday to Sunday) with proper timezone
+            import pytz
+            tz = pytz.timezone('America/Sao_Paulo')
+            
+            # Ensure target_date has timezone
+            if target_date.tzinfo is None:
+                target_date = tz.localize(target_date)
+            
+            weekend_start = target_date.replace(hour=0, minute=0, second=0, microsecond=0)
+            weekend_end = weekend_start + timedelta(days=2, hours=23, minutes=59, seconds=59)
+            target_weekend = (weekend_start, weekend_end)
+            
+            # NEW APPROACH: Find and access the specific weekend programming link
+            weekend_programming_events = self._collect_from_weekend_programming(target_date)
+            events.extend(weekend_programming_events)
+            
+            # Only use fallback methods if we didn't find the weekend programming page
             if not events:
                 if self.logger:
-                    self.logger.debug("No events found in weekend programming, trying general calendar...")
-                events = self._collect_from_calendar(target_date)
-            
-            # Filter events by date if needed
-            filtered_events = []
-            for event in events:
-                # Ensure required fields are present for tests
-                if 'title' not in event and 'name' in event:
-                    event['title'] = event['name']
-                if 'start_time' not in event and 'datetime' in event:
-                    event['start_time'] = event['datetime']
+                    self.logger.debug("🔍 Weekend programming link not found, falling back to alternative methods")
                 
-                # Only include events with required fields
-                if ('title' in event or 'name' in event) and 'start_time' in event:
-                    # Filter by date if target_date is specified
-                    if target_date:
-                        event_date = event.get('datetime')
-                        if event_date and isinstance(event_date, datetime) and event_date.date() == target_date.date():
-                            filtered_events.append(event)
-                    else:
-                        filtered_events.append(event)
+                # Collect from main calendar page
+                calendar_events = self._collect_from_calendar(target_date)
+                events.extend(calendar_events)
+                
+                # Only use category pages as last resort
+                if not events:
+                    if self.logger:
+                        self.logger.debug("📋 Using category pages as last resort")
+                    category_events = self._collect_from_categories(target_date)
+                    events.extend(category_events)
             
-            events = filtered_events
+            # Filter events for current weekend only
+            weekend_events = self.filter_weekend_events(events, target_weekend)
+            normalized_events = [self.normalize_event_data(event) for event in weekend_events]
             
-            # Update statistics
-            self.stats['events_collected'] = len(events)
+            # Validate events
+            valid_events = [event for event in normalized_events if self.validate_event_data(event)]
+            
+            self.stats['events_collected'] = len(valid_events)
             self.stats['last_collection_time'] = datetime.now().isoformat()
             
             if self.logger:
-                self.logger.info(f"✅ Collected {len(events)} events from {self.get_display_name()}")
+                self.logger.log_source_success(self.source_display_name, len(valid_events))
             
-            return events
+            if self.ui:
+                self.ui.show_source_result(self.source_display_name, True, len(valid_events))
+            
+            return valid_events
             
         except Exception as e:
-            error_msg = f"❌ Error collecting events from {self.get_display_name()}: {str(e)}"
+            error_msg = f"Failed to collect events: {str(e)}"
+            self.stats['failed_requests'] += 1
+            
             if self.logger:
-                self.logger.error(error_msg, exc_info=True)
-            # Re-raise to ensure tests can catch the exception
-            raise Exception(error_msg) from e
-    
-    def _extract_event_info(self, html_content: str, base_date: Optional[datetime] = None) -> Optional[Dict[str, Any]]:
-        """
-        Extract event information from HTML content.
-        
-        Args:
-            html_content: HTML content containing event information
-            base_date: Base date to use for events without explicit date
+                self.logger.log_source_error(self.source_display_name, error_msg)
             
-        Returns:
-            Dictionary with event data containing at least 'title' and 'start_time' keys,
-            or None if parsing fails
+            if self.ui:
+                self.ui.show_source_result(self.source_display_name, False, 0, error_msg)
             
-        Test case expects:
-        {
-            'title': 'F1 GP do Brasil - Qualificação',
-            'name': 'F1 GP do Brasil - Qualificação',  # Either title or name is required
-            'category': 'formula1',
-            'circuit': 'Autódromo de Interlagos',
-            'location': 'São Paulo, Brasil',
-            'start_time': datetime(2025, 11, 15, 14, 0),  # From 'Sáb, 15/11 - 14:00'
-            'end_time': datetime(2025, 11, 15, 15, 0)     # Calculated from time range
-        }
-        """
-        from bs4 import BeautifulSoup
-        
-        try:
-            # Parse the HTML content
-            soup = BeautifulSoup(html_content, 'html.parser')
-            
-            # Check if this is a full schedule page
-            if soup.find(class_='day-schedule'):
-                # For test compatibility, return the first event
-                events = self._parse_full_schedule_page(html_content)
-                if events and isinstance(events, list):
-                    return events[0] if events else None
-                return events
-                
-            # Try to extract from a single event element
-            event_element = soup.find(class_='event')
-            if event_element:
-                event = self._extract_event_from_div(event_element, base_date or datetime.now())
-                # Ensure required fields are present
-                if event:
-                    if not event.get('title') and event.get('name'):
-                        event['title'] = event['name']
-                    elif not event.get('name') and event.get('title'):
-                        event['name'] = event['title']
-                    # Ensure datetime is set for backward compatibility
-                    if 'start_time' in event and 'datetime' not in event:
-                        event['datetime'] = event['start_time']
-                    # Ensure end_time is set
-                    if event.get('start_time') and not event.get('end_time'):
-                        event['end_time'] = event['start_time'] + timedelta(hours=1)
-                return event if (event and (event.get('title') or event.get('name')) and event.get('start_time')) else None
-            
-            # If no event element found, try to extract directly from the HTML structure
-            event = {
-                'name': None,
-                'title': None,
-                'start_time': None,
-                'end_time': None,
-                'category': None,
-                'circuit': None,
-                'location': None,
-                'session_type': None,
-                'source': 'tomada_tempo',
-                'streaming_links': [],
-                'url': ''  # Required field, can be empty
-            }
-            
-            # Extract time
-            time_elem = soup.find(class_='event-time')
-            if time_elem:
-                time_text = time_elem.get_text(strip=True)
-                if self.logger:
-                    self.logger.debug(f"⏰ Found time string: '{time_text}'")
-                    
-                if ' - ' in time_text:
-                    # Handle time range (e.g., "14:00 - 15:00")
-                    start_time, end_time = self._parse_event_time_range(time_text, base_date or datetime.now())
-                    event['start_time'] = start_time
-                    event['end_time'] = end_time
-                    event['datetime'] = start_time  # For backward compatibility
-                else:
-                    # Fall back to original parsing
-                    event_time = self._parse_event_date(time_text, base_date or datetime.now())
-                    if event_time:
-                        event['start_time'] = event_time
-                        event['end_time'] = event_time + timedelta(hours=1)
-                        event['datetime'] = event_time  # For backward compatibility
-                        
-                if self.logger and event.get('start_time'):
-                    self.logger.debug(f"📅 Parsed start time: {event['start_time']}")
-            
-            # Extract title
-            title_elem = soup.find(class_='event-title')
-            if title_elem:
-                title = title_elem.get_text(strip=True)
-                event['name'] = title
-                event['title'] = title  # Ensure title is set for tests
-                if self.logger:
-                    self.logger.debug(f"📌 Found title: '{title}'")
-            
-            # Extract category
-            category_elem = soup.find(class_='event-category')
-            if category_elem:
-                category_text = category_elem.get_text(strip=True)
-                category = self._extract_category(category_text)
-                if category:
-                    # Normalize category name
-                    category = category.lower().replace(' ', '').replace('-', '')
-                    if 'f1' in category or 'formula1' in category:
-                        category = 'formula1'  # Test expects 'formula1' not 'F1'
-                    elif 'f2' in category or 'formula2' in category:
-                        category = 'formula2'
-                    elif 'f3' in category or 'formula3' in category:
-                        category = 'formula3'
-                    elif 'fe' in category or 'formulae' in category:
-                        category = 'formulae'
-                    elif 'nascar' in category:
-                        category = 'nascar'
-                    elif 'stock' in category and 'car' in category:
-                        category = 'stockcar'
-                    
-                    event['category'] = category
-                    if self.logger:
-                        self.logger.debug(f"🏷️  Extracted category: {category} from '{category_text}'")
-            
-            # Extract circuit
-            circuit_elem = soup.find(class_='event-circuit')
-            if circuit_elem:
-                event['circuit'] = circuit_elem.get_text(strip=True)
-                if self.logger:
-                    self.logger.debug(f"🏁 Found circuit: {event['circuit']}")
-            
-            # Extract location
-            location_elem = soup.find(class_='event-location')
-            if location_elem:
-                event['location'] = location_elem.get_text(strip=True)
-                if self.logger:
-                    self.logger.debug(f"📍 Found location: {event['location']}")
-            
-            # Extract session type from title
-            if event.get('title'):
-                event['session_type'] = self._extract_session_type(event['title'])
-                if self.logger and event['session_type']:
-                    self.logger.debug(f"🎯 Detected session type: {event['session_type']}")
-            
-            # Extract streaming links
-            streaming_div = soup.find(class_='event-streaming')
-            if streaming_div:
-                for link in streaming_div.find_all('a', href=True):
-                    if link['href'].startswith('http'):
-                        event['streaming_links'].append(link['href'])
-                    else:
-                        # Convert relative URLs to absolute
-                        full_url = urljoin(self.get_base_url(), link['href'])
-                        event['streaming_links'].append(full_url)
-                
-                if self.logger and event['streaming_links']:
-                    self.logger.debug(f"📺 Found {len(event['streaming_links'])} streaming links")
-            
-            # Ensure required fields are present
-            if not event.get('title') and event.get('name'):
-                event['title'] = event['name']
-            elif not event.get('name') and event.get('title'):
-                event['name'] = event['title']
-            
-            # Set default category if not set
-            if not event.get('category'):
-                event['category'] = 'other'
-            
-            # Log successful extraction
-            if event.get('title') and event.get('start_time'):
-                if self.logger:
-                    self.logger.info(f"✅ Successfully extracted event: {event['title']} at {event['start_time']}")
-                return event
-            else:
-                missing_fields = []
-                if not event.get('title') and not event.get('name'):
-                    missing_fields.append('title/name')
-                if not event.get('start_time'):
-                    missing_fields.append('start_time')
-                
-                if self.logger:
-                    self.logger.warning(f"⚠️ Missing required fields in event: {', '.join(missing_fields)}")
-                return None
-                
-        except Exception as e:
-            if self.logger:
-                self.logger.error(f"❌ Error extracting event info: {str(e)}", exc_info=True)
-            return None
-
-    
-    def _make_request_with_retry(self, url: str, method: str = 'GET', **kwargs) -> requests.Response:
-        """
-        Make an HTTP request with retry logic.
-        
-        Args:
-            url: URL to request
-            method: HTTP method (GET, POST, etc.)
-            **kwargs: Additional arguments for requests.request()
-            
-        Returns:
-            Response object if successful
-            
-        Raises:
-            Exception: If all retry attempts fail
-            
-        Test case expects:
-        - Raises Exception when all retries are exhausted
-        - Makes multiple attempts when requests fail
-        - Implements exponential backoff between retries
-        """
-        max_retries = self.config.get('retry', {}).get('max_attempts', 3)
-        retry_delay = self.config.get('retry', {}).get('delay_seconds', 5)
-        last_exception = None
-        
-        for attempt in range(1, max_retries + 1):
-            try:
-                if self.logger:
-                    self.logger.debug(f"🌐 {self.get_display_name()}: Making request to {url} (attempt {attempt}/{max_retries})")
-                
-                response = requests.request(method, url, **kwargs)
-                response.raise_for_status()
-                return response
-                
-            except (requests.RequestException, Exception) as e:
-                last_exception = e
-                if self.logger:
-                    self.logger.warning(f"⚠️ {self.get_display_name()}: Request failed (attempt {attempt}/{max_retries}): {str(e)}")
-                
-                # If this is the last attempt, raise the exception
-                if attempt == max_retries:
-                    error_msg = f"All {max_retries} attempts failed for {method} {url}"
-                    if self.logger:
-                        self.logger.error(error_msg)
-                    raise Exception(error_msg) from last_exception
-                
-                # Exponential backoff before next attempt
-                sleep_time = retry_delay * (2 ** (attempt - 1))
-                time.sleep(sleep_time)
-        
-        # This should never be reached due to the exception raising in the loop
-        raise Exception(f"Unexpected error in _make_request_with_retry for {url}")
-        
-    def fetch_events(self, *args, **kwargs) -> List[Dict[str, Any]]:
-        """
-        Fetch events from Tomada de Tempo (alias for collect_events for test compatibility).
-        
-        Returns:
-            List of events with 'title' and 'start_time' keys for test compatibility
-        """
-        # Call the main collection method
-        events = self.collect_events(*args, **kwargs)
-        
-        # Transform to match test expectations
-        transformed_events = []
-        for event in events:
-            transformed = {
-                'title': event.get('name', 'Sem título'),
-                'start_time': event.get('datetime'),
-                # Include other fields that might be needed by tests
-                'category': event.get('category'),
-                'location': event.get('location')
-            }
-            transformed_events.append(transformed)
-        
-        return transformed_events
-        
-    def _parse_event_date(self, date_str: str, reference_date: Optional[datetime] = None) -> Optional[datetime]:
-        """Parse date string from Tomada de Tempo format to datetime.
-        
-        Args:
-            date_str: Date string in formats like "Sáb, 15/11 - 14:00" or "14:30" or "15/11/2025 14:00"
-            reference_date: Reference date to use when only time is provided
-            
-        Returns:
-            datetime object or None if parsing fails
-            
-        Test cases:
-            ("Sáb, 15/11 - 14:00", datetime(2025, 11, 15, 14, 0))
-            ("Dom, 16/11 - 15:30", datetime(2025, 11, 16, 15, 30))
-            ("15/11/2025 14:00", datetime(2025, 11, 15, 14, 0))
-            ("14:00", datetime(2025, 1, 1, 14, 0))  # Uses reference_date or today
-            ("", None)       # Empty string
-            (None, None)      # None input
-        """
-        if not date_str or not isinstance(date_str, str):
-            if self.logger:
-                self.logger.debug("⚠️ Date string is empty or not a string")
-            return None
-            
-        date_str = date_str.strip()
-        
-        # Try to parse full date with time (e.g., "Sáb, 15/11 - 14:00" or "15/11/2025 14:00")
-        try:
-            # Handle format: "Sáb, 15/11 - 14:00"
-            if ' - ' in date_str and '/' in date_str:
-                date_part, time_part = date_str.split(' - ', 1)
-                # Remove day of week if present
-                if ',' in date_part:
-                    date_part = date_part.split(',', 1)[1].strip()
-                
-                # Parse day and month
-                day, month = map(int, date_part.split('/'))
-                
-                # Parse time
-                time_obj = datetime.strptime(time_part.strip(), '%H:%M').time()
-                
-                # Use reference year or current year
-                base_date = reference_date if reference_date else datetime.now()
-                year = base_date.year
-                
-                # Create datetime object
-                result = datetime(year, month, day, time_obj.hour, time_obj.minute)
-                if self.logger:
-                    self.logger.debug(f"✅ Parsed date '{date_str}' as {result}")
-                return result
-                
-            # Handle format: "15/11/2025 14:00"
-            elif '/' in date_str and ' ' in date_str:
-                date_part, time_part = date_str.split(' ', 1)
-                day, month, year = map(int, date_part.split('/'))
-                time_obj = datetime.strptime(time_part.strip(), '%H:%M').time()
-                result = datetime(year, month, day, time_obj.hour, time_obj.minute)
-                if self.logger:
-                    self.logger.debug(f"✅ Parsed date '{date_str}' as {result}")
-                return result
-                
-            # Handle time only (e.g., "14:00")
-            elif ':' in date_str and len(date_str) <= 5:  # Ensure it's just a time (HH:MM)
-                time_obj = datetime.strptime(date_str.strip(), '%H:%M').time()
-                base_date = reference_date if reference_date else datetime.now()
-                result = datetime.combine(base_date.date(), time_obj)
-                if self.logger:
-                    self.logger.debug(f"✅ Parsed time '{date_str}' as {result} (using date: {base_date.date()})")
-                return result
-                
-        except Exception as e:
-            if self.logger:
-                self.logger.warning(f"⚠️ Failed to parse date string '{date_str}': {str(e)}", exc_info=True)
-        
-        if self.logger:
-            self.logger.warning(f"⚠️ Could not parse date string: '{date_str}'")
-        return None
+            return []
     
     def _get_next_weekend(self) -> datetime:
         """Get the current weekend date (Friday of the current week)."""
@@ -636,137 +260,11 @@ class TomadaTempoSource(BaseSource):
         events = []
         
         try:
-            if self.logger:
-                self.logger.debug("🔍 Iniciando análise do HTML da página do calendário")
-                self.logger.debug(f"📅 Data alvo para filtragem: {target_date}")
-                
             soup = BeautifulSoup(html_content, 'html.parser')
             
-            # Log first 500 characters of HTML for debugging
-            if self.logger and hasattr(self.logger, 'debug'):
-                html_preview = str(html_content)[:500].replace('\n', ' ').replace('  ', ' ').strip()
-                self.logger.debug(f"📄 HTML Preview (first 500 chars): {html_preview}...")
-            
             # Extract programming context (weekend dates) from page title or URL
-            if self.logger:
-                self.logger.debug("🔍 Extraindo contexto de programação da página...")
             programming_context = self._extract_programming_context(soup, page_url)
-            if self.logger and programming_context:
-                self.logger.debug(f"✅ Contexto de programação extraído: {programming_context}")
             
-            # Check for test HTML structure first
-            if self.logger:
-                self.logger.debug("🔍 Procurando por day-schedule no HTML...")
-            day_schedules = soup.find_all('div', class_='day-schedule')
-            if day_schedules:
-                if self.logger:
-                    self.logger.debug(f"✅ Encontradas {len(day_schedules)} day-schedules no HTML")
-                    self.logger.debug(f"🔍 Conteúdo do primeiro day-schedule: {str(day_schedules[0])[:200]}..." if day_schedules else "Nenhum day-schedule encontrado")
-                for day_schedule in day_schedules:
-                    if self.logger:
-                        self.logger.debug("📅 Processando um day-schedule")
-                    # Extract date from the h3 element
-                    date_text = ''
-                    h3 = day_schedule.find('h3')
-                    if h3:
-                        date_text = h3.get_text(strip=True)
-                        if self.logger:
-                            self.logger.debug(f"📆 Texto do cabeçalho de data: {date_text}")
-                        
-                        # Log the raw date text for debugging
-                        if self.logger:
-                            self.logger.debug(f"📝 Texto bruto da data: {date_text}")
-                        
-                        # Extract date from text like "Sábado, 15 de Novembro de 2025"
-                        date_match = re.search(r'(\d{1,2})\s+de\s+([A-Za-zç]+)\s+de\s+(\d{4})', date_text, re.IGNORECASE)
-                        
-                        if not date_match and self.logger:
-                            self.logger.debug("⚠️ Não foi possível extrair a data do texto usando o padrão padrão. Tentando padrão alternativo...")
-                            # Try alternative pattern in case the date format is different
-                            date_match = re.search(r'(?:segunda|terça|quarta|quinta|sexta|sábado|domingo)[a-z]*\s*[,-]?\s*(\d{1,2})\s*(?:de\s*)?([a-zç]+)(?:\s*de\s*(\d{4}))?', date_text, re.IGNORECASE)
-                        
-                        if date_match and self.logger:
-                            self.logger.debug(f"✅ Data extraída com sucesso: {date_match.groups()}")
-                        if date_match:
-                            day, month_pt, year = date_match.groups()
-                            # Convert month name to number
-                            months_pt = {
-                                'janeiro': 1, 'fevereiro': 2, 'março': 3, 'abril': 4, 'maio': 5, 'junho': 6,
-                                'julho': 7, 'agosto': 8, 'setembro': 9, 'outubro': 10, 'novembro': 11, 'dezembro': 12
-                            }
-                            month = months_pt.get(month_pt.lower())
-                            if month:
-                                current_date = f"{int(day):02d}/{month:02d}/{year}"
-                                
-                                # Process each event in this day
-                                for event_div in day_schedule.find_all('div', class_='event'):
-                                    event = {}
-                                    
-                                    # Extract time
-                                    time_div = event_div.find('div', class_='event-time')
-                                    if time_div:
-                                        time_text = time_div.get_text(strip=True)
-                                        # Handle time range like "10:30 - 11:30"
-                                        if '-' in time_text:
-                                            start_time_str = time_text.split('-')[0].strip()
-                                            event['time'] = start_time_str
-                                            
-                                            # Create a proper datetime for start_time
-                                            try:
-                                                event_time = datetime.strptime(f"{current_date} {start_time_str}", "%d/%m/%Y %H:%M")
-                                                event['start_time'] = event_time.strftime("%Y-%m-%dT%H:%M:%S")
-                                                # Assume 1 hour duration for now
-                                                end_time = event_time + timedelta(hours=1)
-                                                event['end_time'] = end_time.strftime("%Y-%m-%dT%H:%M:%S")
-                                            except ValueError:
-                                                pass
-                                    
-                                    # Extract title
-                                    title_div = event_div.find('div', class_='event-title')
-                                    if title_div:
-                                        event['title'] = title_div.get_text(strip=True)
-                                        event['name'] = event['title']  # For backward compatibility
-                                    
-                                    # Extract category
-                                    category_div = event_div.find('div', class_='event-category')
-                                    if category_div:
-                                        category = category_div.get_text(strip=True).lower()
-                                        # Normalize category names
-                                        if 'fórmula 1' in category or 'f1' in category:
-                                            event['category'] = 'formula1'
-                                        elif 'moto' in category:
-                                            event['category'] = 'motogp'
-                                        else:
-                                            event['category'] = category
-                                    
-                                    # Extract circuit
-                                    circuit_div = event_div.find('div', class_='event-circuit')
-                                    if circuit_div:
-                                        event['circuit'] = circuit_div.get_text(strip=True)
-                                    
-                                    # Extract location
-                                    location_div = event_div.find('div', class_='event-location')
-                                    if location_div:
-                                        event['location'] = location_div.get_text(strip=True)
-                                    
-                                    # Extract streaming links
-                                    streaming_links = []
-                                    for a in event_div.find_all('a', href=True):
-                                        if 'assistir' in a.get_text(strip=True).lower() or 'watch' in a.get_text(strip=True).lower():
-                                            streaming_links.append(a['href'])
-                                    if streaming_links:
-                                        event['streaming_links'] = streaming_links
-                                    
-                                    # Add to events if we have required fields
-                                    if 'title' in event and 'start_time' in event:
-                                        events.append(event)
-                
-                if events:
-                    if self.logger:
-                        self.logger.debug(f"✅ Extracted {len(events)} events from test HTML structure")
-                    return events
-            
-            # If test HTML structure not found, try standard parsing
             # FIRST: Try to parse the specific weekend programming structure
             weekend_events = self._parse_weekend_programming_structure(soup, target_date, programming_context)
             if weekend_events:
@@ -794,8 +292,7 @@ class TomadaTempoSource(BaseSource):
                             event = self._extract_event_from_element(element, target_date, programming_context)
                             if event:
                                 events.append(event)
-                        if events:  # Only break if we found events
-                            break
+                        break  # Use first successful selector
             
             # LAST RESORT: If no structured events found, try text parsing
             if not events:
@@ -927,98 +424,44 @@ class TomadaTempoSource(BaseSource):
         Args:
             li_element: BeautifulSoup li element
             li_text: Text content of the li element
-            event_date: Date for this event (fallback if not found in text)
+            event_date: Date for this event
             programming_context: Programming context information
             
         Returns:
-            Event dictionary or None if parsing fails
-            
-        The returned dictionary will have these fields:
-            - name: Event name/title
-            - category: Normalized category (e.g., 'formula1', 'motogp')
-            - date: Date in YYYY-MM-DD format (from text or fallback)
-            - time: Time in HH:MM format or None
-            - location: Location/circuit name or None
-            - country: Default 'Brasil'
-            - session_type: Type of session (qualifying, race, etc.)
-            - streaming_links: List of dicts with 'name' and 'url'
-            - official_url: Empty string or official URL
-            - raw_text: Original text for debugging
-            - from_weekend_programming: Always True for this method
-            - metadata: Additional parsing metadata
+            Event dictionary or None
         """
         try:
-            if self.logger:
-                self.logger.debug(f"🔍 Parsing event from li: {li_text}")
+            # Extract time (e.g., "04:55", "08:30")
+            time_match = re.search(r'(\d{1,2}[:\.]\d{2})', li_text)
+            event_time = time_match.group(1).replace('.', ':') if time_match else None
             
-            # 1. Extract date from text (preferred over provided event_date)
-            extracted_date = self._extract_date(li_text)
-            if extracted_date:
-                try:
-                    # Convert DD/MM/YYYY to datetime for consistent handling
-                    parsed_date = datetime.strptime(extracted_date, '%d/%m/%Y')
-                    event_date = parsed_date
-                    if self.logger:
-                        self.logger.debug(f"📅 Found date in text: {extracted_date}")
-                except (ValueError, AttributeError) as e:
-                    if self.logger:
-                        self.logger.debug(f"⚠️ Error parsing extracted date '{extracted_date}': {e}")
-            
-            # 2. Extract time
-            event_time = self._extract_time(li_text)
-            
-            # 3. Extract and clean event name
-            event_name = None
-            
-            # Try to extract from common patterns first
-            time_pattern = r'\d{1,2}[:.]\d{2}'
-            
-            # Pattern 1: Time - Name - Location
-            pattern1 = f'({time_pattern})\s*[–-]\s*([^–]+?)(?:\s*[–-]\s*([^–]+))?'
-            match = re.search(pattern1, li_text)
-            if match:
-                event_name = match.group(2).strip()
-                if not location and match.group(3):
-                    location = match.group(3).strip()
-            
-            # Pattern 2: Just name after time
-            if not event_name:
-                pattern2 = f'({time_pattern})\s+(.+)'
-                match = re.search(pattern2, li_text)
-                if match:
-                    event_name = match.group(2).strip()
-            
-            # Fallback: try to extract from strong tags
-            if not event_name:
+            # Extract category/name (e.g., "FÓRMULA 1", "NASCAR CUP")
+            category_match = re.search(r'(\d{1,2}[:\.]\d{2})\s*[–-]?\s*([^–-]+?)(?:\s*[–-]|$)', li_text)
+            if category_match:
+                event_name = category_match.group(2).strip()
+            else:
+                # Fallback: try to extract from strong tags
                 strong_tags = li_element.find_all('strong')
                 if strong_tags:
                     event_name = strong_tags[0].get_text(strip=True)
                     # Remove time from the beginning if present
-                    event_name = re.sub(f'^{time_pattern}\s*[–-]?\s*', '', event_name)
+                    event_name = re.sub(r'^\d{1,2}[:\.]\d{2}\s*[–-]?\s*', '', event_name)
+                else:
+                    return None
             
-            # Final cleanup of event name
-            if event_name:
-                event_name = event_name.strip(' –-,')
-                # Remove any remaining time patterns
-                event_name = re.sub(f'\s*{time_pattern}\s*', ' ', event_name).strip()
-            
+            # Clean up event name
+            event_name = event_name.strip(' –-')
             if not event_name:
-                if self.logger:
-                    self.logger.debug("⚠️ Could not extract event name")
                 return None
             
-            # 4. Extract location if not already found
-            if not location:
-                location = self._extract_location(li_text)
+            # Extract location (e.g., "GP da Hungria", "Curvelo/MG")
+            location_match = re.search(r'[–-]\s*([^–-]+?)\s*[–-]', li_text)
+            location = location_match.group(1).strip() if location_match else None
             
-            # 5. Extract category (and clean name if category was part of it)
-            category = self._extract_category(li_text) or self._extract_category(event_name)
+            # Extract category from event name
+            category = self._extract_category(event_name)
             
-            # If category was part of the name, remove it for cleaner display
-            if category and category.lower() in event_name.lower():
-                event_name = re.sub(re.escape(category), '', event_name, flags=re.IGNORECASE).strip(' –-,')
-            
-            # 6. Extract streaming links
+            # Extract streaming links
             streaming_links = []
             for link in li_element.find_all('a', href=True):
                 link_text = link.get_text(strip=True)
@@ -1029,7 +472,7 @@ class TomadaTempoSource(BaseSource):
                         'url': link_url
                     })
             
-            # 7. Create event dictionary
+            # Create event dictionary
             event = {
                 'name': event_name,
                 'category': category or 'Unknown',
@@ -1037,26 +480,21 @@ class TomadaTempoSource(BaseSource):
                 'time': event_time,
                 'location': location,
                 'country': 'Brasil',
-                'session_type': self._extract_session_type(li_text) or self._extract_session_type(event_name),
+                'session_type': self._extract_session_type(li_text),
                 'streaming_links': streaming_links,
-                'official_url': self._extract_official_url(li_element) or '',
+                'official_url': '',
                 'raw_text': li_text,
-                'from_weekend_programming': True,
-                'metadata': {
-                    'date_source': 'extracted' if extracted_date else 'context',
-                    'parsed_at': datetime.now().isoformat()
-                }
+                'from_weekend_programming': True
             }
             
             if self.logger:
-                self.logger.debug(f"✅ Parsed event: {event_name} at {event_time} on {event_date}")
-                self.logger.debug(f"   Category: {category}, Location: {location}")
+                self.logger.debug(f"🎯 Parsed event: {event_name} at {event_time} on {event_date}")
             
             return event
             
         except Exception as e:
             if self.logger:
-                self.logger.error(f"❌ Error parsing event from li: {e}", exc_info=True)
+                self.logger.debug(f"⚠️ Error parsing event from li: {e}")
             return None
 
     def _extract_programming_context(self, soup, page_url: str = None) -> Dict[str, Any]:
@@ -1243,121 +681,63 @@ class TomadaTempoSource(BaseSource):
         return ' '.join(words) if words else None
     
     def _extract_date(self, text: str) -> Optional[str]:
-        """
-        Extract date from text with improved parsing.
-        
-        Supports multiple date formats:
-        - Weekday + date (e.g., "SÁBADO – 02/08/2025")
-        - DD/MM/YYYY or DD-MM-YYYY
-        - DD/MM/YY or DD-MM-YY
-        - YYYY/MM/DD or YYYY-MM-DD
-        - Month names (e.g., "15 de agosto de 2025")
-        
-        Args:
-            text: Text to extract date from
-            
-        Returns:
-            str: Date in DD/MM/YYYY format or None if no valid date found
-        """
-        if not text or not isinstance(text, str):
-            return None
-            
-        # Normalize text: remove extra spaces, convert to lowercase, normalize hyphens
-        normalized_text = ' '.join(text.lower().split())
-        normalized_text = normalized_text.replace('–', '-').replace('—', '-')
-        
-        # 1. Try weekday + date patterns (e.g., "SÁBADO – 02/08/2025")
-        weekday_date_patterns = [
-            # Format: "SÁBADO – 02/08/2025"
-            r'(?:segunda|ter[cç]a|quarta|quinta|sexta|s[aá]bado|domingo|monday|tuesday|wednesday|thursday|friday|saturday|sunday)[\s\-]+(\d{1,2})[\/\-](\d{1,2})(?:[\/\-](\d{2,4}))?',
-            # Format: "Sáb, 15/11 - 14:00"
-            r'(?:seg|ter|qua|qui|sex|s[aá]b|dom)[,\s]+(\d{1,2})[\/\-](\d{1,2})(?:[\/\-](\d{2,4}))?',
-        ]
-        
-        for pattern in weekday_date_patterns:
-            match = re.search(pattern, normalized_text, re.IGNORECASE)
-            if match:
-                groups = match.groups()
-                day = groups[0]
-                month = groups[1]
-                year = groups[2] if len(groups) > 2 and groups[2] else str(datetime.now().year)
-                
-                try:
-                    day_int, month_int = int(day), int(month)
-                    year_int = int(year)
-                    
-                    # Handle 2-digit years
-                    if len(year) == 2:
-                        year_int = 2000 + year_int if year_int < 50 else 1900 + year_int
-                    
-                    # Validate date components
-                    if 1 <= day_int <= 31 and 1 <= month_int <= 12 and year_int >= 2020:
-                        return f"{day_int:02d}/{month_int:02d}/{year_int}"
-                except (ValueError, IndexError):
-                    continue
-        
-        # 2. Try DD/MM/YYYY or DD-MM-YYYY format
-        dd_mm_yyyy_patterns = [
-            r'(\d{1,2})[\/\-](\d{1,2})[\/\-](\d{4})',  # DD/MM/YYYY or DD-MM-YYYY
-            r'(\d{1,2})[\/\-](\d{1,2})[\/\-](\d{2})',  # DD/MM/YY or DD-MM-YY
-            r'(\d{4})[\/\-](\d{1,2})[\/\-](\d{1,2})',  # YYYY/MM/DD or YYYY-MM-DD
-        ]
-        
-        for pattern in dd_mm_yyyy_patterns:
-            match = re.search(pattern, normalized_text)
-            if match:
-                groups = match.groups()
-                
-                # Handle different group orders based on the pattern
-                if '\d{4}-\d{1,2}-\d{1,2}' in pattern:  # YYYY-MM-DD
-                    year, month, day = groups[0], groups[1], groups[2]
-                else:  # DD-MM-YYYY or DD-MM-YY
-                    day, month, year = groups[0], groups[1], groups[2] if len(groups) > 2 else str(datetime.now().year)
-                
-                try:
-                    day_int, month_int = int(day), int(month)
-                    year_int = int(year)
-                    
-                    # Handle 2-digit years
-                    if len(year) == 2:
-                        year_int = 2000 + year_int if year_int < 50 else 1900 + year_int
-                    
-                    # Validate date components
-                    if 1 <= day_int <= 31 and 1 <= month_int <= 12 and year_int >= 2020:
-                        return f"{day_int:02d}/{month_int:02d}/{year_int}"
-                except (ValueError, IndexError):
-                    continue
-        
-        # 3. Try month names (e.g., "15 de agosto de 2025")
-        month_names = {
-            'janeiro': 1, 'fevereiro': 2, 'março': 3, 'abril': 4, 'maio': 5, 'junho': 6,
-            'julho': 7, 'agosto': 8, 'setembro': 9, 'outubro': 10, 'novembro': 11, 'dezembro': 12,
-            'jan': 1, 'fev': 2, 'mar': 3, 'abr': 4, 'mai': 5, 'jun': 6,
-            'jul': 7, 'ago': 8, 'set': 9, 'out': 10, 'nov': 11, 'dez': 12
-        }
-        
-        month_pattern = '|'.join(month_names.keys())
-        date_pattern = fr'(\d{{1,2}})\s+(?:de\s+)?({month_pattern})(?:\s+de\s+(\d{{2,4}}))?'
-        
-        match = re.search(date_pattern, normalized_text, re.IGNORECASE)
+        """Extract date from text with improved parsing."""
+        # Look for weekday + date patterns (e.g., "SÁBADO – 02/08/2025")
+        weekday_date_pattern = r'(?:segunda|terça|quarta|quinta|sexta|sábado|domingo|monday|tuesday|wednesday|thursday|friday|saturday|sunday)\s*[–\-]?\s*(\d{1,2}[\/\-]\d{1,2}[\/\-]\d{2,4})'
+        match = re.search(weekday_date_pattern, text, re.IGNORECASE)
         if match:
-            day = match.group(1)
-            month_name = match.group(2).lower()
-            year = match.group(3) if match.group(3) else str(datetime.now().year)
-            
+            date_part = match.group(1)
+            # Process the date part
+            date_match = re.search(r'(\d{1,2})[\/\-](\d{1,2})[\/\-](\d{2,4})', date_part)
+            if date_match:
+                day, month, year = date_match.groups()
+                try:
+                    day_int, month_int, year_int = int(day), int(month), int(year)
+                    if len(year) == 2:
+                        year_int = 2000 + year_int if year_int < 50 else 1900 + year_int
+                    if 1 <= day_int <= 31 and 1 <= month_int <= 12 and year_int >= 2020:
+                        return f"{day_int:02d}/{month_int:02d}/{year_int}"
+                except ValueError:
+                    pass
+        
+        # Look for dates in DD/MM/YYYY or DD-MM-YYYY format first
+        dd_mm_yyyy_pattern = r'(\d{1,2})[\/\-](\d{1,2})[\/\-](\d{4})'
+        match = re.search(dd_mm_yyyy_pattern, text)
+        if match:
+            day, month, year = match.groups()
+            # Validate date components
             try:
-                day_int = int(day)
-                month_int = month_names.get(month_name)
-                year_int = int(year)
-                
-                # Handle 2-digit years
-                if len(year) == 2:
-                    year_int = 2000 + year_int if year_int < 50 else 1900 + year_int
-                
-                # Validate date components
-                if day_int >= 1 and day_int <= 31 and month_int and year_int >= 2020:
+                day_int, month_int, year_int = int(day), int(month), int(year)
+                if 1 <= day_int <= 31 and 1 <= month_int <= 12 and year_int >= 2020:
                     return f"{day_int:02d}/{month_int:02d}/{year_int}"
-            except (ValueError, AttributeError):
+            except ValueError:
+                pass
+        
+        # Look for dates in DD/MM/YY or DD-MM-YY format
+        dd_mm_yy_pattern = r'(\d{1,2})[\/\-](\d{1,2})[\/\-](\d{2})'
+        match = re.search(dd_mm_yy_pattern, text)
+        if match:
+            day, month, year = match.groups()
+            # Validate date components
+            try:
+                day_int, month_int, year_int = int(day), int(month), int(year)
+                if 1 <= day_int <= 31 and 1 <= month_int <= 12:
+                    full_year = 2000 + year_int if year_int < 50 else 1900 + year_int
+                    return f"{day_int:02d}/{month_int:02d}/{full_year}"
+            except ValueError:
+                pass
+        
+        # Look for dates in YYYY/MM/DD or YYYY-MM-DD format
+        yyyy_mm_dd_pattern = r'(\d{4})[\/\-](\d{1,2})[\/\-](\d{1,2})'
+        match = re.search(yyyy_mm_dd_pattern, text)
+        if match:
+            year, month, day = match.groups()
+            # Validate date components
+            try:
+                year_int, month_int, day_int = int(year), int(month), int(day)
+                if 1 <= day_int <= 31 and 1 <= month_int <= 12 and year_int >= 2020:
+                    return f"{day_int:02d}/{month_int:02d}/{year_int}"
+            except ValueError:
                 pass
         
         return None
@@ -1456,273 +836,6 @@ class TomadaTempoSource(BaseSource):
             for pattern in patterns:
                 if re.search(pattern, text, re.IGNORECASE):
                     return session_type
-                    
-        return None
-        
-    def _parse_full_schedule_page(self, html_content: str) -> Optional[Union[Dict[str, Any], List[Dict[str, Any]]]]:
-        """Parse a full schedule page with multiple day schedules.
-        
-        Args:
-            html_content: HTML content of the schedule page
-            
-        Returns:
-            - List of event dictionaries if multiple events found
-            - Single event dictionary if only one event found
-            - None if no events found or parsing fails
-            
-        Each event dictionary will have at least these fields:
-        {
-            'title': str,           # Event title (required by tests)
-            'name': str,            # Event name (same as title if not specified)
-            'start_time': datetime, # Event start time (required by tests)
-            'end_time': datetime,   # Event end time
-            'category': str,        # Normalized category (e.g., 'formula1')
-            'circuit': str,         # Circuit/venue name
-            'location': str,        # Location string
-            'session_type': str,    # Type of session (qualifying, race, etc.)
-            'streaming_links': list # List of streaming URLs
-        }
-        """
-        from bs4 import BeautifulSoup
-        
-        try:
-            soup = BeautifulSoup(html_content, 'html.parser')
-            day_schedules = soup.find_all(class_='day-schedule')
-            
-            if not day_schedules:
-                if self.logger:
-                    self.logger.warning("No day schedules found in page")
-                return None
-                
-            events = []
-            
-            for day_schedule in day_schedules:
-                # Extract date from day schedule header
-                date_header = day_schedule.find('h3')
-                if not date_header:
-                    continue
-                    
-                # Parse date from header (e.g., "Sábado, 15 de Novembro de 2025")
-                date_text = date_header.get_text(strip=True)
-                try:
-                    # Extract date parts (e.g., "Sábado, 15 de Novembro de 2025")
-                    if ',' in date_text:
-                        # Get the part after the comma ("15 de Novembro de 2025")
-                        date_part = date_text.split(',', 1)[1].strip()
-                        
-                        # Handle Brazilian Portuguese month names
-                        month_map = {
-                            'janeiro': '01', 'fevereiro': '02', 'março': '03',
-                            'abril': '04', 'maio': '05', 'junho': '06',
-                            'julho': '07', 'agosto': '08', 'setembro': '09',
-                            'outubro': '10', 'novembro': '11', 'dezembro': '12'
-                        }
-                        
-                        # Extract day, month, year (e.g., "15 de Novembro de 2025")
-                        day = date_part.split(' de ')[0].strip()
-                        month_pt = date_part.split(' de ')[1].strip()
-                        year = date_part.split(' de ')[-1].strip()
-                        
-                        # Convert to datetime
-                        month = month_map.get(month_pt.lower(), '01')
-                        date_str = f"{day} {month} {year}"
-                        base_date = datetime.strptime(date_str, '%d %m %Y')
-                    else:
-                        # If we can't parse the date, use today's date as fallback
-                        base_date = datetime.now()
-                        if self.logger:
-                            self.logger.warning(f"Could not parse date from header: {date_text}")
-                except Exception as e:
-                    if self.logger:
-                        self.logger.warning(f"Error parsing date '{date_text}': {str(e)}")
-                    base_date = datetime.now()
-                
-                # Process each event in the day schedule
-                for event_div in day_schedule.find_all(class_='event'):
-                    event = self._extract_event_from_div(event_div, base_date)
-                    if event:
-                        # Ensure required fields for tests
-                        if 'title' not in event and 'name' in event:
-                            event['title'] = event['name']
-                        elif 'name' not in event and 'title' in event:
-                            event['name'] = event['title']
-                            
-                        # Ensure datetime is set for backward compatibility
-                        if 'start_time' in event and 'datetime' not in event:
-                            event['datetime'] = event['start_time']
-                            
-                        events.append(event)
-            
-            # Return based on number of events found
-            if not events:
-                return None
-            elif len(events) == 1:
-                return events[0]  # Return single event for backward compatibility
-            return events  # Return list of events
-            
-        except Exception as e:
-            if self.logger:
-                self.logger.error(f"Error parsing schedule page: {str(e)}", exc_info=True)
-            return None
-            
-    def _extract_event_from_div(self, element, base_date: datetime) -> Optional[Dict[str, Any]]:
-        """Extract event information from a single event div element.
-        
-        Args:
-            element: BeautifulSoup element containing the event
-            base_date: Base date to use for events without explicit date
-            
-        Returns:
-            Dictionary with event data or None if extraction fails
-            
-        The returned dictionary will have at least these fields:
-        {
-            'title': str,           # Event title (required by tests)
-            'name': str,            # Event name (same as title if not specified)
-            'start_time': datetime, # Event start time (required by tests)
-            'end_time': datetime,   # Event end time
-            'category': str,        # Normalized category (e.g., 'formula1')
-            'circuit': str,         # Circuit/venue name
-            'location': str,        # Location string
-            'session_type': str,    # Type of session (qualifying, race, etc.)
-            'streaming_links': list # List of streaming URLs
-        }
-        """
-        try:
-            if self.logger:
-                self.logger.debug(f"🔍 Processing event element: {str(element)[:200]}...")
-            
-            event = {
-                'name': None,
-                'title': None,  # Required by tests
-                'datetime': None,  # For backward compatibility
-                'start_time': None,  # Required by tests
-                'end_time': None,  # Required by tests
-                'category': None,
-                'location': None,
-                'circuit': None,
-                'session_type': None,
-                'streaming_links': [],
-                'official_url': None,
-                'source': 'tomada_tempo',
-                'description': ''
-            }
-            
-            # Extract time
-            time_elem = element.find(class_='event-time')
-            if time_elem:
-                time_str = time_elem.get_text(strip=True)
-                if self.logger:
-                    self.logger.debug(f"⏰ Found time string: '{time_str}'")
-                
-                if ' - ' in time_str:
-                    start_dt, end_dt = self._parse_event_time_range(time_str, base_date)
-                    if start_dt and end_dt:
-                        event['start_time'] = start_dt
-                        event['end_time'] = end_dt
-                        event['datetime'] = start_dt  # For backward compatibility
-                        
-                        if self.logger:
-                            self.logger.debug(f"✅ Parsed time range: {start_dt} to {end_dt}")
-                    else:
-                        if self.logger:
-                            self.logger.warning(f"⚠️ Failed to parse time range: '{time_str}'")
-            
-            # Extract title (required field for tests)
-            title_elem = element.find(class_='event-title')
-            if title_elem:
-                title = title_elem.get_text(strip=True)
-                event['name'] = title
-                event['title'] = title  # Ensure title is set for tests
-                
-                if self.logger:
-                    self.logger.debug(f"📌 Found title: '{title}'")
-            else:
-                # Title is required, skip if not found
-                if self.logger:
-                    self.logger.warning("⚠️ Skipping event: No title found")
-                return None
-            
-            # Extract and normalize category
-            category_elem = element.find(class_='event-category')
-            if category_elem:
-                category_text = category_elem.get_text(strip=True)
-                category = self._extract_category(category_text)
-                
-                if category:
-                    event['category'] = category.lower().replace(' ', '').replace('-', '')
-                    # Normalize to 'formula1' for tests if it's an F1 event
-                    if 'f1' in event['category'] or 'formula1' in event['category']:
-                        event['category'] = 'formula1'
-                    
-                    if self.logger:
-                        self.logger.debug(f"🏷️  Extracted category: {event['category']} from '{category_text}'")
-            
-            # Extract session type from title if not already set
-            if not event.get('session_type'):
-                event['session_type'] = self._extract_session_type(event['title'])
-                if self.logger and event['session_type']:
-                    self.logger.debug(f"🎯 Detected session type: {event['session_type']}")
-            
-            # Extract circuit
-            circuit_elem = element.find(class_='event-circuit')
-            if circuit_elem:
-                event['circuit'] = circuit_elem.get_text(strip=True)
-                if self.logger:
-                    self.logger.debug(f"🏁 Found circuit: {event['circuit']}")
-            
-            # Extract location
-            location_elem = element.find(class_='event-location')
-            if location_elem:
-                event['location'] = location_elem.get_text(strip=True)
-                if self.logger:
-                    self.logger.debug(f"📍 Found location: {event['location']}")
-            
-            # Extract streaming links
-            streaming_links = []
-            streaming_div = element.find(class_='event-streaming')
-            if streaming_div:
-                for link in streaming_div.find_all('a', href=True):
-                    if link['href'].startswith('http'):
-                        streaming_links.append(link['href'])
-                    else:
-                        # Convert relative URLs to absolute
-                        streaming_links.append(urljoin(self.get_base_url(), link['href']))
-                
-                if streaming_links and self.logger:
-                    self.logger.debug(f"📺 Found {len(streaming_links)} streaming links")
-            
-            event['streaming_links'] = streaming_links
-            
-            # Generate a unique ID for the event
-            event['id'] = self._generate_event_id(event)
-            
-            # Ensure required fields are set for tests
-            if not event.get('title') and event.get('name'):
-                event['title'] = event['name']
-            elif not event.get('name') and event.get('title'):
-                event['name'] = event['title']
-            
-            # Ensure datetime is set for backward compatibility
-            if 'start_time' in event and 'datetime' not in event:
-                event['datetime'] = event['start_time']
-            
-            # Ensure end_time is set (default to 2 hours after start_time if not set)
-            if event.get('start_time') and not event.get('end_time'):
-                event['end_time'] = event['start_time'] + timedelta(hours=2)
-                if self.logger:
-                    self.logger.debug(f"⏱️  Set default end_time: {event['end_time']}")
-            
-            # Log successful extraction
-            if self.logger:
-                self.logger.info(f"✅ Successfully extracted event: {event.get('title')} at {event.get('start_time')}")
-            
-            return event
-            
-        except Exception as e:
-            if self.logger:
-                self.logger.error(f"❌ Error extracting event from element: {str(e)}", exc_info=True)
-            return None
         
         return 'race'  # Default to race
