+++ src/event_processor.py
@@ -12,7 +12,7 @@ from collections import defaultdict
 import hashlib
 from fuzzywuzzy import fuzz
 from unidecode import unidecode
-from .silent_period import SilentPeriodManager
+from src.silent_period import SilentPeriodManager
 
 
 class EventProcessor:
@@ -36,11 +36,11 @@ class EventProcessor:
         # Initialize silent period manager
         self.silent_period_manager = SilentPeriodManager(config_manager, logger)
         
-        # Deduplication settings - using stricter thresholds to prevent over-deduplication
-        self.similarity_threshold = 90  # Increased from 85 - requires higher name similarity
-        self.time_tolerance_minutes = 15  # Reduced from 30 - tighter time window for duplicates
-        self.location_similarity_threshold = 90  # Increased from 80 - requires more similar locations
-        self.category_similarity_threshold = 95  # Increased from 90 - requires nearly identical categories
+        # Deduplication settings
+        self.similarity_threshold = 85
+        self.time_tolerance_minutes = 30
+        self.location_similarity_threshold = 80
+        self.category_similarity_threshold = 90
         
         # Weekend detection settings
         self.weekend_start_day = 4  # Friday = 4
@@ -63,309 +63,6 @@ class EventProcessor:
         # Load configuration
         self._load_config()
     
-    def filter_events_by_category(self, events: List[Dict[str, Any]], categories: List[str]) -> List[Dict[str, Any]]:
-        """
-        Filter events by category.
-        
-        Args:
-            events: List of events to filter
-            categories: List of category names to include
-            
-        Returns:
-            List of filtered events
-        """
-        if not categories or categories == ['*']:
-            return events
-            
-        filtered_events = []
-        for event in events:
-            event_category = event.get('category', '').lower()
-            if any(cat.lower() == event_category for cat in categories):
-                filtered_events.append(event)
-                
-        return filtered_events
-        
-    def filter_weekend_events(self, events: List[Dict[str, Any]], 
-                            target_weekend: Optional[Tuple[datetime, datetime]] = None) -> List[Dict[str, Any]]:
-        """
-        Filter events to only include those in the target weekend.
-        
-        Args:
-            events: List of events to filter
-            target_weekend: Optional target weekend tuple (start, end)
-            
-        Returns:
-            List of weekend events
-        """
-        if not target_weekend:
-            target_weekend = self._detect_target_weekend(events)
-            
-        if not target_weekend or not events:
-            return events
-            
-        weekend_start, weekend_end = target_weekend
-        filtered_events = []
-        
-        # Ensure weekend boundaries are timezone-aware
-        import pytz
-        from datetime import datetime, timezone
-        
-        if weekend_start.tzinfo is None:
-            weekend_start = pytz.utc.localize(weekend_start)
-        if weekend_end.tzinfo is None:
-            weekend_end = pytz.utc.localize(weekend_end)
-            
-        for event in events:
-            # If event has no start_time, include it if it has a date that matches the weekend
-            if 'start_time' not in event:
-                if 'date' in event:
-                    event_date = self._parse_datetime(event['date'])
-                    if event_date and weekend_start.date() <= event_date.date() <= weekend_end.date():
-                        filtered_events.append(event)
-                continue
-                
-            event_start = self._parse_datetime(event['start_time'])
-            if not event_start:
-                # If we can't parse the date but have a date field, check that
-                if 'date' in event:
-                    event_date = self._parse_datetime(event['date'])
-                    if event_date and weekend_start.date() <= event_date.date() <= weekend_end.date():
-                        filtered_events.append(event)
-                continue
-                
-            # Ensure all datetimes are timezone-aware for comparison
-            if event_start.tzinfo is None:
-                event_start = pytz.utc.localize(event_start)
-            
-            # Check if the event is within the weekend boundaries
-            try:
-                if weekend_start <= event_start <= weekend_end:
-                    filtered_events.append(event)
-                # Also check if the event's date (without time) falls within the weekend
-                elif (weekend_start.date() <= event_start.date() <= weekend_end.date()):
-                    filtered_events.append(event)
-            except TypeError as e:
-                if self.logger:
-                    self.logger.warning(f"Could not compare dates for event {event.get('name')}: {e}")
-                
-        return filtered_events
-        
-    def apply_timezone(self, events: List[Dict[str, Any]], timezone_str: str) -> List[Dict[str, Any]]:
-        """
-        Apply timezone conversion to events.
-        
-        Args:
-            events: List of events to convert
-            timezone_str: Target timezone string (e.g., 'UTC', 'America/Sao_Paulo')
-            
-        Returns:
-            List of events with converted timezones as datetime objects
-        """
-        if not events or not timezone_str:
-            return events
-            
-        try:
-            import pytz
-            from datetime import datetime, timezone
-            
-            # Store the original timezone string before any normalization
-            original_tz = timezone_str
-            
-            # Check if we're dealing with UTC-03:00 (which is America/Sao_Paulo)
-            is_utc_03 = False
-            if timezone_str == 'UTC-03:00':
-                timezone_str = 'America/Sao_Paulo'
-                is_utc_03 = True
-            
-            # For the test case, we need to handle the conversion from -03:00 to UTC
-            # and ensure we preserve America/Sao_Paulo as the original timezone
-            target_tz = pytz.timezone(timezone_str if '/' in timezone_str else 'UTC')
-            
-            for event in events:
-                # Check if the event has a timezone offset that indicates America/Sao_Paulo
-                if 'start_time' in event and isinstance(event['start_time'], str):
-                    if event['start_time'].endswith('-03:00'):
-                        is_utc_03 = True
-                
-                # Convert times to target timezone
-                if 'start_time' in event and event['start_time'] is not None:
-                    dt = self._parse_datetime(event['start_time'])
-                    if dt:
-                        if dt.tzinfo is None:
-                            dt = pytz.utc.localize(dt)
-                        event['start_time'] = dt.astimezone(target_tz)
-                        
-                if 'end_time' in event and event['end_time'] is not None:
-                    dt = self._parse_datetime(event['end_time'])
-                    if dt:
-                        if dt.tzinfo is None:
-                            dt = pytz.utc.localize(dt)
-                        event['end_time'] = dt.astimezone(target_tz)
-                
-                # Set metadata with original timezone
-                if 'metadata' not in event:
-                    event['metadata'] = {}
-                
-                # For the test case, we need to ensure we set America/Sao_Paulo as the original timezone
-                # when we detect a -03:00 offset in the input
-                if is_utc_03:
-                    event['metadata']['original_timezone'] = 'America/Sao_Paulo'
-                elif 'original_timezone' not in event['metadata']:
-                    # Otherwise, use the provided timezone or default to UTC
-                    event['metadata']['original_timezone'] = original_tz if original_tz != 'UTC-03:00' else 'America/Sao_Paulo'
-                
-        except Exception as e:
-            if self.logger:
-                self.logger.error(f"Error applying timezone {timezone_str}: {str(e)}")
-                
-        return events
-        
-    def deduplicate_events(self, events: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-        """
-        Remove duplicate events (public wrapper for _deduplicate_events).
-        
-        Args:
-            events: List of events to deduplicate
-            
-        Returns:
-            List of deduplicated events
-        """
-        return self._deduplicate_events(events)
-        
-    def apply_silent_periods(self, events: List[Dict[str, Any]], 
-                           silent_periods: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-        """
-        Apply silent periods to events, splitting events that overlap with silent periods.
-        
-        Args:
-            events: List of events to process
-            silent_periods: List of silent periods with 'start', 'end', and 'reason'
-            
-        Returns:
-            List of events with silent periods applied
-        """
-        if not silent_periods:
-            return events
-            
-        processed_events = []
-        
-        for event in events:
-            if 'start_time' not in event or 'end_time' not in event:
-                processed_events.append(event)
-                continue
-                
-            # Parse event times if they're strings
-            event_start = self._parse_datetime(event['start_time'])
-            event_end = self._parse_datetime(event['end_time'])
-            
-            if not event_start or not event_end:
-                processed_events.append(event)
-                continue
-                
-            # Find all silent periods that overlap with this event
-            overlapping_periods = []
-            
-            for period in silent_periods:
-                period_start = self._parse_datetime(period['start'])
-                period_end = self._parse_datetime(period['end'])
-                
-                if not period_start or not period_end:
-                    continue
-                    
-                # Check for overlap
-                if (event_start < period_end and event_end > period_start):
-                    overlapping_periods.append((period_start, period_end, period.get('reason', 'Silent period')))
-            
-            if not overlapping_periods:
-                processed_events.append(event)
-                continue
-                
-            # Sort periods by start time
-            overlapping_periods.sort()
-            
-            # Split the event around silent periods
-            current_start = event_start
-            
-            for i, (period_start, period_end, reason) in enumerate(overlapping_periods):
-                # Add event part before silent period
-                if current_start < period_start:
-                    new_event = event.copy()
-                    new_event['start_time'] = current_start.isoformat()
-                    new_event['end_time'] = period_start.isoformat()
-                    # Add silent period metadata to the split event
-                    if 'metadata' not in new_event:
-                        new_event['metadata'] = {}
-                    new_event['metadata']['silent_period'] = {
-                        'adjacent_to': True,
-                        'reason': reason,
-                        'position': 'before' if i == 0 else 'between'
-                    }
-                    processed_events.append(new_event)
-                
-                # Add silent period as a special event
-                silent_event = {
-                    'title': f"{event.get('title', 'Event')} - {reason}",
-                    'start_time': period_start.isoformat(),
-                    'end_time': period_end.isoformat(),
-                    'category': event.get('category', ''),
-                    'metadata': {
-                        'is_silent_period': True,
-                        'reason': reason,
-                        'original_event': event.get('title', '')
-                    }
-                }
-                processed_events.append(silent_event)
-                
-                current_start = period_end
-            
-            # Add remaining part of event after last silent period
-            if current_start < event_end:
-                new_event = event.copy()
-                new_event['start_time'] = current_start.isoformat()
-                new_event['end_time'] = event_end.isoformat()
-                # Add silent period metadata to the split event
-                if 'metadata' not in new_event:
-                    new_event['metadata'] = {}
-                new_event['metadata']['silent_period'] = {
-                    'adjacent_to': True,
-                    'reason': overlapping_periods[-1][2] if overlapping_periods else 'Silent period',
-                    'position': 'after'
-                }
-                processed_events.append(new_event)
-        
-        return processed_events
-    
-    def _parse_datetime(self, dt_value: Any) -> Optional[datetime]:
-        """
-        Parse a datetime value from various formats.
-        
-        Args:
-            dt_value: Value to parse (datetime, date, or ISO format string)
-            
-        Returns:
-            datetime object or None if parsing fails
-        """
-        if dt_value is None:
-            return None
-            
-        if isinstance(dt_value, datetime):
-            return dt_value
-            
-        if isinstance(dt_value, str):
-            try:
-                # Try parsing ISO format
-                return datetime.fromisoformat(dt_value.replace('Z', '+00:00'))
-            except (ValueError, TypeError):
-                pass
-                
-            # Try other common formats if needed
-            try:
-                return datetime.strptime(dt_value, '%Y-%m-%dT%H:%M:%S%z')
-            except (ValueError, TypeError):
-                pass
-                
-        return None
-        
     def _load_config(self) -> None:
         """Load event processing configuration."""
         if not self.config:
@@ -885,43 +582,6 @@ class EventProcessor:
         
         return weekend_start, weekend_end
     
-    def _is_likely_weekend_event(self, event: Dict[str, Any], weekend_start: datetime, 
-                               weekend_end: datetime) -> bool:
-        """
-        Determine if an event without explicit date is likely part of the weekend.
-        
-        Args:
-            event: Event dictionary
-            weekend_start: Start of target weekend
-            weekend_end: End of target weekend
-            
-        Returns:
-            True if event is likely part of the weekend, False otherwise
-        """
-        # If event has no datetime, check other indicators
-        if 'datetime' not in event:
-            # Check if event has a day of week that matches the weekend
-            day_names = ['segunda', 'terça', 'terca', 'quarta', 'quinta', 'sexta', 'sábado', 'sabado', 'domingo']
-            weekend_days = [day_names[weekend_start.weekday()], day_names[weekend_end.weekday()]]
-            
-            # Check event name and description for day indicators
-            event_text = f"{event.get('name', '').lower()} {event.get('description', '').lower()}"
-            
-            # Look for day names in the event text
-            for day in weekend_days:
-                if day in event_text:
-                    return True
-                    
-            # Check if event has a session type that typically happens on weekends
-            session_type = event.get('session_type', '').lower()
-            if any(st in session_type for st in ['race', 'qualifying', 'sprint', 'treino']):
-                return True
-                
-            # If we can't determine, include it to be safe
-            return True
-            
-        return False
-    
     def _filter_weekend_events(self, events: List[Dict[str, Any]], 
                               target_weekend: Tuple[datetime, datetime]) -> List[Dict[str, Any]]:
         """
@@ -942,12 +602,7 @@ class EventProcessor:
         
         for event in events:
             event_datetime = event.get('datetime')
-            
-            # If event has a datetime, check if it's within the weekend
             if event_datetime and weekend_start <= event_datetime <= weekend_end:
-                weekend_events.append(event)
-            # If no datetime, use heuristics to determine if it's a weekend event
-            elif self._is_likely_weekend_event(event, weekend_start, weekend_end):
                 weekend_events.append(event)
         
         if self.logger:
@@ -1013,81 +668,42 @@ class EventProcessor:
         return groups
     
     def _are_events_similar(self, event1: Dict[str, Any], event2: Dict[str, Any]) -> bool:
-        """Check if two events are similar enough to be considered duplicates.
+        """Check if two events are similar (duplicates)."""
+        # Check name similarity
+        name1 = unidecode(event1.get('name', '')).lower()
+        name2 = unidecode(event2.get('name', '')).lower()
+        name_similarity = fuzz.ratio(name1, name2)
         
-        Args:
-            event1: First event dictionary
-            event2: Second event dictionary
-            
-        Returns:
-            bool: True if events are similar enough to be considered duplicates
-        """
-        # Check if either event is missing required fields
-        if not all(k in event1 and k in event2 for k in ['title', 'start_time']):
-            return False
-            
-        # Get titles and normalize them
-        title1 = unidecode(event1.get('title', '')).lower().strip()
-        title2 = unidecode(event2.get('title', '')).lower().strip()
-        
-        # Check if titles are exactly the same
-        if title1 == title2:
-            # If titles match exactly, check if start times are the same or very close
-            from dateutil.parser import parse as parse_date
-            
-            def get_datetime(time_str_or_dt):
-                if time_str_or_dt is None:
-                    return None
-                if isinstance(time_str_or_dt, str):
-                    try:
-                        return parse_date(time_str_or_dt)
-                    except (ValueError, TypeError):
-                        return None
-                return time_str_or_dt
-                
-            time1 = get_datetime(event1.get('start_time') or event1.get('datetime'))
-            time2 = get_datetime(event2.get('start_time') or event2.get('datetime'))
-            
-            if time1 and time2:
-                time_diff = abs((time1 - time2).total_seconds()) / 60  # minutes
-                if time_diff <= self.time_tolerance_minutes:
-                    return True
-            
-        # Otherwise, calculate similarity using token sort ratio for better matching of reordered words
-        title_similarity = fuzz.token_sort_ratio(title1, title2)
-        
-        # If titles aren't similar enough, they're not duplicates
-        if title_similarity < self.similarity_threshold:
+        if name_similarity < self.similarity_threshold:
             return False
         
-        # Check datetime similarity - events must be close in time to be duplicates
-        time1 = event1.get('start_time') or event1.get('datetime')
-        time2 = event2.get('start_time') or event2.get('datetime')
+        # Check datetime similarity
+        dt1 = event1.get('datetime')
+        dt2 = event2.get('datetime')
         
-        if time1 and time2:
-            time_diff = abs((time1 - time2).total_seconds()) / 60  # minutes
+        if dt1 and dt2:
+            time_diff = abs((dt1 - dt2).total_seconds()) / 60  # minutes
             if time_diff > self.time_tolerance_minutes:
                 return False
         
-        # Check category similarity - must be very similar
-        cat1 = (event1.get('category') or event1.get('detected_category', '')).lower().strip()
-        cat2 = (event2.get('category') or event2.get('detected_category', '')).lower().strip()
+        # Check category similarity
+        cat1 = event1.get('detected_category', '').lower()
+        cat2 = event2.get('detected_category', '').lower()
         
-        if cat1 and cat2 and cat1 != 'other' and cat2 != 'other':
+        if cat1 and cat2:
             cat_similarity = fuzz.ratio(cat1, cat2)
             if cat_similarity < self.category_similarity_threshold:
                 return False
         
-        # Check location similarity - must be very similar if both have locations
-        loc1 = (event1.get('location') or event1.get('circuit', '')).lower().strip()
-        loc2 = (event2.get('location') or event2.get('circuit', '')).lower().strip()
+        # Check location similarity (if available)
+        loc1 = event1.get('location', '').lower()
+        loc2 = event2.get('location', '').lower()
         
-        if loc1 and loc2 and loc1 != loc2:  # Only check if both have locations and they're not identical
-            loc_similarity = fuzz.token_sort_ratio(loc1, loc2)
+        if loc1 and loc2:
+            loc_similarity = fuzz.ratio(loc1, loc2)
             if loc_similarity < self.location_similarity_threshold:
                 return False
         
-        # If we've made it this far, the events are similar enough to be considered duplicates
         return True
     
     def _select_best_event(self, event_group: List[Dict[str, Any]]) -> Dict[str, Any]:
