--- src/sources/sources
@@ -1,1 +1,0 @@
-/Users/daniel.m.santos/My Drive/Projetos/Race Calendar/sources
