--- tests/__init__.py
@@ -1,7 +1,0 @@
-"""
-Test package for the Motorsport Calendar application.
-This file makes the tests directory a Python package, which helps with imports.
-"""
-
-# This file is intentionally left empty.
-# It marks the tests directory as a Python package.
