--- tests/config/test_config.json
@@ -1,191 +1,0 @@
-{
-  "sources": [
-    {
-      "name": "tomada_tempo",
-      "enabled": true,
-      "base_url": "https://www.tomadatempo.com.br",
-      "endpoints": {
-        "programacao": "/programacao",
-        "evento": "/evento",
-        "categorias": "/categorias"
-      },
-      "timeout": 5,
-      "retry_attempts": 2,
-      "headers": {
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
-      }
-    }
-  ],
-  
-  "output": {
-    "directory": "./test_output",
-    "filename_format": "events_{date}.ics",
-    "date_format": "%Y%m%d",
-    "overwrite_existing": true
-  },
-  
-  "logging": {
-    "level": "INFO",
-    "file_structure": {
-      "main_log": "test_logs/test.log",
-      "debug_directory": "test_logs/debug",
-      "payload_directory": "test_logs/payloads"
-    },
-    "retention": {
-      "enabled": true,
-      "max_logs_to_keep": 3,
-      "max_payloads_to_keep": 5,
-      "delete_older_than_days": 1
-    },
-    "levels": {
-      "console": "INFO",
-      "file": "DEBUG",
-      "debug_file": "DEBUG"
-    },
-    "format": {
-      "console": "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-      "file": "%(asctime)s - %(name)s - %(levelname)s - %(funcName)s:%(lineno)d - %(message)s"
-    },
-    "rotation": {
-      "enabled": true,
-      "max_size_mb": 1,
-      "backup_count": 2
-    },
-    "payload_settings": {
-      "save_raw": true,
-      "pretty_print": false,
-      "include_headers": false,
-      "separate_by_source": true
-    }
-  },
-  
-  "retry": {
-    "max_attempts": 3,
-    "backoff_factor": 1,
-    "status_forcelist": [500, 502, 503, 504]
-  },
-  
-  "ui": {
-    "enabled": true,
-    "theme": "dark",
-    "refresh_interval_seconds": 300,
-    "show_progress_bars": true
-  },
-  
-  "calendar": {
-    "timezone": "UTC",
-    "default_duration_minutes": 120,
-    "reminders": [
-      {
-        "minutes": 60,
-        "method": "popup"
-      },
-      {
-        "minutes": 15,
-        "method": "popup"
-      }
-    ]
-  },
-  
-  "filters": {
-    "include_categories": ["formula1", "formula2", "formula3", "nascar", "wec", "indycar", "motogp"],
-    "exclude_categories": [],
-    "min_duration_minutes": 30,
-    "max_duration_hours": 6
-  },
-  
-  "event_processing": {
-    "timezone_handling": "convert_to_utc",
-    "duplicate_detection": {
-      "enabled": true,
-      "window_minutes": 30,
-      "check_title_similarity": true,
-      "check_datetime_proximity": true,
-      "similarity_threshold": 0.8
-    },
-    "date_parsing": {
-      "formats": [
-        "%Y-%m-%dT%H:%M:%S%z",
-        "%Y-%m-%d %H:%M:%S",
-        "%d/%m/%Y %H:%M",
-        "%d/%m/%Y %Hh%M"
-      ],
-      "default_timezone": "UTC",
-      "prefer_locale": true,
-      "fallback_timezone": "America/Sao_Paulo"
-    },
-    "category_detection": {
-      "enabled": true,
-      "confidence_threshold": 0.7,
-      "fallback_category": "other"
-    }
-  },
-  
-  "general": {
-    "timezone": "UTC",
-    "language": "en-US",
-    "log_level": "INFO",
-    "output_directory": "./test_output",
-    "visual_interface": {
-      "enabled": true,
-      "colors": true,
-      "progress_bars": true,
-      "icons": true
-    },
-    "advanced_logging": {
-      "debug_logs": true,
-      "payload_storage": true,
-      "log_rotation": true,
-      "retention_days": 7
-    },
-    "silent_periods": []
-  },
-  
-  "event_filters": {
-    "included_categories": ["*"],
-    "excluded_categories": [],
-    "included_countries": [],
-    "excluded_countries": [],
-    "event_types": {
-      "race": true,
-      "qualifying": true,
-      "practice": true,
-      "sprint": true,
-      "other": false
-    },
-    "min_duration_minutes": 30,
-    "max_duration_days": 7
-  },
-  
-  "ical_parameters": {
-    "calendar_name": "Test Motorsport Calendar",
-    "calendar_description": "Test calendar for motorsport events",
-    "timezone": "UTC",
-    "default_duration_minutes": 120,
-    "reminders": [
-      {
-        "minutes": 60,
-        "method": "popup"
-      },
-      {
-        "minutes": 15,
-        "method": "popup"
-      }
-    ],
-    "description_template": "{title}\\n\\nLocal: {location}\\nCategoria: {category}\\n\\n{description}",
-    "include_streaming_links": true
-  },
-  
-
-  
-  "advanced": {
-    "enable_debug_endpoints": false,
-    "log_payloads": false,
-    "cache_ttl_seconds": 3600,
-    "validate_responses": true,
-    "max_events_per_request": 1000
-  },
-  
-  "version": "1.0.0",
-  "environment": "test"
-}
