--- tests/regression/test_data/sample_events.json
@@ -1,71 +1,0 @@
-[
-  {
-    "source": "tomada_tempo",
-    "title": "F1 GP do Brasil - Treino Livre 1",
-    "start_time": "2025-11-15T10:30:00-03:00",
-    "end_time": "2025-11-15T11:30:00-03:00",
-    "category": "formula1",
-    "circuit": "Autódromo de Interlagos",
-    "location": "São Paulo, Brasil",
-    "description": "Primeiro treino livre do GP do Brasil de Fórmula 1",
-    "streaming_links": ["https://example.com/stream1"],
-    "metadata": {
-      "series": "Fórmula 1",
-      "session_type": "practice",
-      "round": 21,
-      "season": 2025
-    }
-  },
-  {
-    "source": "tomada_tempo",
-    "title": "F1 GP do Brasil - Qualificação",
-    "start_time": "2025-11-15T14:00:00-03:00",
-    "end_time": "2025-11-15T15:00:00-03:00",
-    "category": "formula1",
-    "circuit": "Autódromo de Interlagos",
-    "location": "São Paulo, Brasil",
-    "description": "Sessão de classificação do GP do Brasil de Fórmula 1",
-    "streaming_links": ["https://example.com/stream1"],
-    "metadata": {
-      "series": "Fórmula 1",
-      "session_type": "qualifying",
-      "round": 21,
-      "season": 2025
-    }
-  },
-  {
-    "source": "tomada_tempo",
-    "title": "F1 GP do Brasil - Corrida",
-    "start_time": "2025-11-16T14:00:00-03:00",
-    "end_time": "2025-11-16T16:00:00-03:00",
-    "category": "formula1",
-    "circuit": "Autódromo de Interlagos",
-    "location": "São Paulo, Brasil",
-    "description": "Corrida do GP do Brasil de Fórmula 1",
-    "streaming_links": ["https://example.com/stream1"],
-    "metadata": {
-      "series": "Fórmula 1",
-      "session_type": "race",
-      "round": 21,
-      "season": 2025,
-      "laps": 71,
-      "distance": "305.909 km"
-    }
-  },
-  {
-    "source": "ergast_api",
-    "title": "F1 GP de Abu Dhabi - Treino Livre 1",
-    "start_time": "2025-12-05T10:30:00+04:00",
-    "end_time": "2025-12-05T11:30:00+04:00",
-    "category": "formula1",
-    "circuit": "Yas Marina Circuit",
-    "location": "Abu Dhabi, UAE",
-    "description": "First practice session for the Abu Dhabi Grand Prix",
-    "metadata": {
-      "series": "Formula 1",
-      "session_type": "practice",
-      "round": 22,
-      "season": 2025
-    }
-  }
-]
