--- tests/regression/test_data/test_config.json
@@ -1,88 +1,0 @@
-{
-  "general": {
-    "timezone": "America/Sao_Paulo",
-    "language": "pt-BR",
-    "log_level": "INFO",
-    "output_directory": "./output",
-    "visual_interface": {
-      "enabled": true,
-      "colors": true,
-      "progress_bars": true,
-      "icons": true
-    },
-    "advanced_logging": {
-      "debug_logs": true,
-      "payload_storage": true,
-      "log_rotation": true,
-      "retention_days": 7
-    },
-    "silent_periods": []
-  },
-  "data_sources": {
-    "priority_order": ["tomada_tempo", "ergast_api"],
-    "excluded_sources": [],
-    "timeout_seconds": 5,
-    "retry_attempts": 2,
-    "rate_limit_delay": 0.5
-  },
-  "event_filters": {
-    "category_detection": {
-      "enabled": true,
-      "learning_mode": false,
-      "confidence_threshold": 0.7
-    },
-    "included_categories": ["*"],
-    "excluded_categories": [],
-    "included_countries": [],
-    "excluded_countries": [],
-    "event_types": {
-      "race": true,
-      "qualifying": true,
-      "practice": false,
-      "sprint": true,
-      "other": false
-    }
-  },
-  "ical_parameters": {
-    "calendar_name": "Test Motorsport Events",
-    "calendar_description": "Test calendar for regression testing",
-    "timezone": "America/Sao_Paulo",
-    "default_duration_minutes": 120,
-    "reminders": [
-      {"minutes": 60, "method": "popup"}
-    ]
-  },
-  "logging": {
-    "file_structure": {
-      "main_log": "test.log",
-      "debug_directory": "debug",
-      "payload_directory": "payloads"
-    },
-    "retention": {
-      "enabled": true,
-      "max_logs_to_keep": 3,
-      "max_payloads_to_keep": 5,
-      "delete_older_than_days": 1
-    },
-    "levels": {
-      "console": "INFO",
-      "file": "DEBUG",
-      "debug_file": "DEBUG"
-    },
-    "format": {
-      "console": "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-      "file": "%(asctime)s - %(name)s - %(levelname)s - %(funcName)s:%(lineno)d - %(message)s"
-    },
-    "rotation": {
-      "enabled": true,
-      "max_size_mb": 1,
-      "backup_count": 2
-    },
-    "payload_settings": {
-      "save_raw": true,
-      "pretty_print": false,
-      "include_headers": false,
-      "separate_by_source": true
-    }
-  }
-}
