--- tests/regression/test_data/tomada_tempo_sample.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Programação de Automobilismo</title>
-</head>
-<body>
-    <div class="schedule-container">
-        <div class="day-schedule">
-            <h3>Sábado, 15 de Novembro de 2025</h3>
-            <div class="event">
-                <div class="event-time">10:30 - 11:30</div>
-                <div class="event-title">F1 GP do Brasil - Treino Livre 1</div>
-                <div class="event-category">Fórmula 1</div>
-                <div class="event-circuit">Autódromo de Interlagos</div>
-                <div class="event-location">São Paulo, Brasil</div>
-                <div class="event-streaming">
-                    <a href="https://example.com/stream1" target="_blank">Assistir</a>
-                </div>
-            </div>
-            
-            <div class="event">
-                <div class="event-time">14:00 - 15:00</div>
-                <div class="event-title">F1 GP do Brasil - Qualificação</div>
-                <div class="event-category">Fórmula 1</div>
-                <div class="event-circuit">Autódromo de Interlagos</div>
-                <div class="event-location">São Paulo, Brasil</div>
-                <div class="event-streaming">
-                    <a href="https://example.com/stream1" target="_blank">Assistir</a>
-                </div>
-            </div>
-        </div>
-        
-        <div class="day-schedule">
-            <h3>Domingo, 16 de Novembro de 2025</h3>
-            <div class="event">
-                <div class="event-time">14:00 - 16:00</div>
-                <div class="event-title">F1 GP do Brasil - Corrida</div>
-                <div class="event-category">Fórmula 1</div>
-                <div class="event-circuit">Autódromo de Interlagos</div>
-                <div class="event-location">São Paulo, Brasil</div>
-                <div class="event-metadata">
-                    <span class="laps">71 voltas</span>
-                    <span class="distance">305.909 km</span>
-                </div>
-                <div class="event-streaming">
-                    <a href="https://example.com/stream1" target="_blank">Assistir</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
