--- tests/regression/test_files/test_file_manager.py
@@ -1,41 +1,0 @@
-"""
-Regression tests for file management functionality.
-
-Note: This test module is currently a placeholder as the FileManager class
-has not been implemented yet. The tests are marked with xfail to indicate
-expected failures until the implementation is complete.
-"""
-import pytest
-from datetime import datetime
-
-@pytest.mark.xfail(reason="FileManager implementation pending")
-class TestFileManager:
-    """Test cases for FileManager (implementation pending)."""
-    
-    def test_ensure_directories(self, file_manager, tmp_path):
-        """Test that required directories are created (not implemented)."""
-        assert False, "FileManager implementation pending"
-    
-    def test_cleanup_old_files(self, file_manager, tmp_path):
-        """Test cleanup of old files (not implemented)."""
-        assert False, "FileManager implementation pending"
-    
-    def test_rotate_logs(self, file_manager, tmp_path):
-        """Test log rotation functionality (not implemented)."""
-        assert False, "FileManager implementation pending"
-    
-    def test_save_payload(self, file_manager, tmp_path):
-        """Test saving of payload files (not implemented)."""
-        assert False, "FileManager implementation pending"
-    
-    def test_cleanup_payloads(self, file_manager, tmp_path):
-        """Test cleanup of old payload files (not implemented)."""
-        assert False, "FileManager implementation pending"
-    
-    def test_save_and_load_events(self, file_manager, tmp_path):
-        """Test saving and loading events (not implemented)."""
-        assert False, "FileManager implementation pending"
-    
-    def test_handle_file_errors(self, file_manager, tmp_path):
-        """Test error handling for file operations (not implemented)."""
-        assert False, "FileManager implementation pending"
